+++ CHANGELOG.md
@@ -51,6 +51,8 @@ this format is adapted from [Keep a Changelog](https://keepachangelog.com/en/1.1
 - Added the option to use logarithmic scales for the length × coverage
   scatterplots.
 
+- Added some basic pattern label options (`Offset` and `Outline`).
+
 - You can now select and copy text from the node/edge/pattern/path tables
   ([#298](https://github.com/marbl/MetagenomeScope/issues/298)).